* V5-7-patches:
  MinGW: Make sure that the backslashes in the pskill message are displayed.
  Compiler warning fix for systems where uint64_t = unsigned long (e.g. RHEL 6.1/x86_64)

+++ agent/mibgroup/hardware/fsys/hw_fsys.c
@@ -342,6 +342,6 @@ netsnmp_fsys_calculate32(netsnmp_fsys_info *f)
     DEBUGMSGTL(("fsys", "Results of 32-bit conversion: size %" PRIu64 " -> %lu;"
 		" units %" PRIu64 " -> %lu; avail %" PRIu64 " -> %lu;"
                 " used %" PRIu64 " -> %lu\n",
-		f->size, f->size_32, f->units, f->units_32,
-		f->avail, f->avail_32, f->used, f->used_32));
+		(uint64_t)f->size, f->size_32, (uint64_t)f->units, f->units_32,
+		(uint64_t)f->avail, f->avail_32, (uint64_t)f->used, f->used_32));
 }
